เปลี่ยน PostgreSQL ใช้ CommonJS and ลบ login  ย้ายไปรวมกับ users
